+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 0.1.2b2 (unreleased)
 
 
-- Nothing changed yet.
+- Added hooks for field and model validation
 
 
 ## 0.1.2.beta1 (2023-08-22)